Stripped templates from main analysis builder script
