+++ config/templates/analysis_cfg.py
@@ -1,0 +1,40 @@
+#!/usr/bin/env python3
+
+#Optional: List of analysis packages to load in runtime
+analysesList = ['__ANALYSIS_NAME__']
+
+#Mandatory: List of processes
+processList = {
+    'p8_noBES_ee_H_Hbb_ecm125':{'fraction':0.01, 'chunks':1, 'output':'test_out'}
+}
+
+#Mandatory: Production tag when running over EDM4Hep centrally produced events, this points to the yaml files for getting sample statistics
+prodTag     = "FCCee/spring2021/IDEA/"
+
+#Optional: output directory, default is local running directory
+outputDir   = "."
+
+#Optional
+nCPUS       = 8
+runBatch    = False
+#batchQueue = "longlunch"
+#compGroup = "group_u_FCC.local_gen"
+
+#Optional test file
+testFile ="root://eospublic.cern.ch//eos/experiment/fcc/ee/generation/DelphesEvents/spring2021/IDEA/p8_ee_ZH_ecm240/events_101027117.root"
+
+#Mandatory: RDFanalysis class where the use defines the operations on the TTree
+class RDFanalysis():
+    #__________________________________________________________
+    #Mandatory: analysers function to define the analysers to process, please make sure you return the last dataframe, in this example it is df2
+    def analysers(df):
+        df2 = (df
+               .Define("dummy_collection", "__ANALYSIS_NAME__::dummy_collection(ReconstructedParticles)")
+              )
+        return df2
+
+    #__________________________________________________________
+    #Mandatory: output function, please make sure you return the branchlist as a python list
+    def output():
+        branchList = ['dummy_collection']
+        return branchList
