+++ config/templates/classes.h
@@ -1,0 +1,3 @@
+namespace __ANALYSIS_NAME__ {
+  struct dictionary {};
+}  // namespace __ANALYSIS_NAME__
