+++ config/templates/classes_def.xml
@@ -1,0 +1,3 @@
+<lcgdict>
+  <class name="__ANALYSIS_NAME__::dictionary"/>
+</lcgdict>
